requirements.txt
flask==3.0.2
anthropic==0.17.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
pygments==2.17.2
markdown==3.5.2
python-markdown-math==0.8